fix: 716 marge dashboard 🐛
