+++ apps/web/components/Layout/DashboardHeader/index.tsx
@@ -15,12 +15,16 @@ import {
   Left,
   LogoButton,
   ChannelText,
+  Nav,
+  NavItem,
+  ProfileButton,
 } from "./styles";
 import { MonoText } from "@/components/UI/Monotext";
 import { useTranslation } from "react-i18next";
 import Link from "next/link";
 import styled from "styled-components";
 import { PATHS } from "@/utils/path";
+import { ROLE_CREATOR, ROLE_VIEWER } from "@/utils/Constants";
 import {
   getLoginUserEmail,
   getLoginUserInitial,
@@ -29,12 +33,13 @@ import {
 import { useCreatorChannelLayout } from "@/hooks/useCreatorChannelLayout";
 
 type Props = {
+  role: typeof ROLE_CREATOR | typeof ROLE_VIEWER;
   onToggleSidebar?: () => void;
+  onProfileClick?: () => void;
 };
 
-const DashboardHeader = ({ onToggleSidebar }: Props) => {
+const DashboardHeader = ({ role, onToggleSidebar, onProfileClick }: Props) => {
   const { t } = useTranslation();
-  const { channelHref } = useCreatorChannelLayout();
   const user = useStoredLoginUser();
   const email = getLoginUserEmail(user);
   const initial = getLoginUserInitial(user);
@@ -66,26 +71,27 @@ const DashboardHeader = ({ onToggleSidebar }: Props) => {
         </LogoButton>
       </Left>
 
+      {role === ROLE_VIEWER && (
+        <Nav>
+          <NavItem href={PATHS.HOW_IT_WORKS}>{t("nav.howItWorks")}</NavItem>
+          <NavItem href={PATHS.EXPLORE_CREATORS}>
+            {t("nav.exploreCreators")}
+          </NavItem>
+          <NavItem href={PATHS.ABOUT}>{t("nav.about")}</NavItem>
+        </Nav>
+      )}
+
       <Right>
-        <ChannelLink href={channelHref}>
-          <ChannelText $use="Body_Medium">
-            {t("dashboard.creatorHeader.myChannel")}
-          </ChannelText>
-        </ChannelLink>
-        <Divider />
-        <Link
-          href={PATHS.DASHBOARD_CREATOR_PROFILE}
-          aria-label={t("common.creatorProfile")}
-        >
-          <RightProfileWrapper>
-            <ProfileCircle>
-              <InitialAvatar>{initial}</InitialAvatar>
-            </ProfileCircle>
-            <EmailWrapper>
-              <MonoText $use="Body_Medium">{email}</MonoText>
-            </EmailWrapper>
-          </RightProfileWrapper>
-        </Link>
+        {role === ROLE_CREATOR ? (
+          <CreatorHeaderRight initial={initial} email={email} />
+        ) : (
+          <ProfileButton
+            aria-label={t("common.viewerProfile")}
+            onClick={onProfileClick}
+          >
+            {initial || "V"}
+          </ProfileButton>
+        )}
       </Right>
     </HeaderWrapper>
   );
@@ -107,5 +113,40 @@ const RightProfileWrapper = styled.div`
   cursor: pointer;
   text-decoration: none;
 `;
+
+const CreatorHeaderRight = ({
+  initial,
+  email,
+}: {
+  initial: string;
+  email: string;
+}) => {
+  const { t } = useTranslation();
+  const { channelHref } = useCreatorChannelLayout();
+
+  return (
+    <>
+      <ChannelLink href={channelHref}>
+        <ChannelText $use="Body_Medium">
+          {t("dashboard.creatorHeader.myChannel")}
+        </ChannelText>
+      </ChannelLink>
+      <Divider />
+      <Link
+        href={PATHS.DASHBOARD_CREATOR_PROFILE}
+        aria-label={t("common.creatorProfile")}
+      >
+        <RightProfileWrapper>
+          <ProfileCircle>
+            <InitialAvatar>{initial}</InitialAvatar>
+          </ProfileCircle>
+          <EmailWrapper>
+            <MonoText $use="Body_Medium">{email}</MonoText>
+          </EmailWrapper>
+        </RightProfileWrapper>
+      </Link>
+    </>
+  );
+};
 
 export default DashboardHeader;
