+++ apps/web/utils/Constants.ts
@@ -38,6 +38,8 @@ export type PasswordVisibilityKey =
 
 export const STRING = "string";
 export const VIEW = "view";
+export const ROLE_CREATOR = "creator";
+export const ROLE_VIEWER = "viewer";
 export const HTTP_STATUS_UNAUTHORIZED = 401;
 export const WEBSITE = "website";
 export const OPEN_GRAPH_LOCALE_EN_US = "en_US";
